テストの作成のためにmainから一部の関数を移動しました．
